# [gql_code_builder] optimize slow code generation to reduce build time as 10% of previous.