--- a/codegen/gql_build/lib/src/utils/config.dart
+++ b/codegen/gql_build/lib/src/utils/config.dart
@@ -42,6 +42,10 @@ EnumFallbackConfig enumFallbackConfig(Map<String, dynamic> config) =>
       fallbackValueMap: enumFallbackMap(config["enum_fallbacks"]),
     );
 
+DataClassConfig dataClassConfig(Map<String, dynamic> config) => DataClassConfig(
+      reuseFragments: config["reuse_fragments"] == true,
+    );
+
 InlineFragmentSpreadWhenExtensionConfig whenExtensionConfig(
     Map<String, dynamic> config) {
   final whenYamlConfig = config["when_extensions"] as YamlMap?;