--- a/codegen/gql_code_builder/lib/src/built_class.dart
+++ b/codegen/gql_code_builder/lib/src/built_class.dart
@@ -11,6 +11,7 @@ Class builtClass({
   Map<String, Expression>? initializers,
   Map<String, SourceSelections> superclassSelections = const {},
   List<Method> methods = const [],
+  Map<String, Reference>? dataClassAliasMap,
 }) {
   final className = builtClassName(name);
   return Class(
@@ -30,7 +31,9 @@ Class builtClass({
                 ],
               ),
           ),
-          ...superclassSelections.keys.map<Reference>(
+          ...superclassSelections.keys
+              .where((superName) => dataClassAliasMap?.containsKey(builtClassName(superName)) != true)
+              .map<Reference>(
             (superName) => refer(
               builtClassName(superName),
               (superclassSelections[superName]?.url ?? "") + "#data",