--- a/codegen/gql_code_builder/lib/src/config/data_class_config.dart
+++ b/codegen/gql_code_builder/lib/src/config/data_class_config.dart
@@ -1,0 +1,8 @@
+/// config for the optimization of data class generation.
+class DataClassConfig {
+  final bool reuseFragments;
+
+  const DataClassConfig({
+    required this.reuseFragments,
+  });
+}