--- a/codegen/gql_code_builder/lib/src/inline_fragment_classes.dart
+++ b/codegen/gql_code_builder/lib/src/inline_fragment_classes.dart
@@ -22,6 +22,7 @@ List<Spec> buildInlineFragmentClasses({
   required String type,
   required Map<String, Reference> typeOverrides,
   required Map<String, SourceSelections> fragmentMap,
+  required Map<String, Reference> dataClassAliasMap,
   required Map<String, SourceSelections> superclassSelections,
   required List<InlineFragmentNode> inlineFragments,
   required bool built,
@@ -31,6 +32,7 @@ List<Spec> buildInlineFragmentClasses({
     baseTypeName: name,
     inlineFragments: inlineFragments,
     config: whenExtensionConfig,
+    dataClassAliasMap: dataClassAliasMap,
   );
   return [
     Class(
@@ -38,12 +40,14 @@ List<Spec> buildInlineFragmentClasses({
         ..abstract = true
         ..name = builtClassName(name)
         ..implements.addAll(
-          superclassSelections.keys.map<Reference>(
-            (superName) => refer(
-              builtClassName(superName),
-              (superclassSelections[superName]?.url ?? "") + "#data",
+          superclassSelections.keys
+            .where((superName) => !dataClassAliasMap.containsKey(builtClassName(superName)))
+            .map<Reference>(
+              (superName) => refer(
+                builtClassName(superName),
+                (superclassSelections[superName]?.url ?? "") + "#data",
+              ),
             ),
-          ),
         )
         ..methods.addAll([
           ...fieldGetters,
@@ -51,6 +55,7 @@ List<Spec> buildInlineFragmentClasses({
             ..._inlineFragmentRootSerializationMethods(
               name: builtClassName(name),
               inlineFragments: inlineFragments,
+              dataClassAliasMap: dataClassAliasMap,
             ),
         ]),
     ),
@@ -65,6 +70,7 @@ List<Spec> buildInlineFragmentClasses({
         fragmentMap,
       ),
       fragmentMap: fragmentMap,
+      dataClassAliasMap: dataClassAliasMap,
       schemaSource: schemaSource,
       type: type,
       typeOverrides: typeOverrides,
@@ -77,7 +83,17 @@ List<Spec> buildInlineFragmentClasses({
 
     /// TODO: Handle inline fragments without a type condition
     /// https://spec.graphql.org/June2018/#sec-Inline-Fragments
-    ...inlineFragments.where((frag) => frag.typeCondition != null).expand(
+    ...inlineFragments.where((frag) {
+      if (frag.typeCondition == null) {
+        return false;
+      }
+      final typeName = builtClassName("${name}__as${frag.typeCondition!.on.name.value}");
+      if (dataClassAliasMap.containsKey(typeName)) {
+        // print("alias $typeName => ${dataClassAliasMap[typeName]!.symbol}");
+        return false;
+      }
+      return true;
+    }).expand(
           (inlineFragment) => buildSelectionSetDataClasses(
               name: "${name}__as${inlineFragment.typeCondition!.on.name.value}",
               selections: mergeSelections(
@@ -89,6 +105,7 @@ List<Spec> buildInlineFragmentClasses({
                 fragmentMap,
               ),
               fragmentMap: fragmentMap,
+              dataClassAliasMap: dataClassAliasMap,
               schemaSource: schemaSource,
               type: inlineFragment.typeCondition!.on.name.value,
               typeOverrides: typeOverrides,
@@ -104,6 +121,7 @@ List<Spec> buildInlineFragmentClasses({
 List<Method> _inlineFragmentRootSerializationMethods({
   required String name,
   required List<InlineFragmentNode> inlineFragments,
+  required Map<String, Reference> dataClassAliasMap,
 }) =>
     [
       buildSerializerGetter(name).rebuild(
@@ -121,7 +139,7 @@ List<Method> _inlineFragmentRootSerializationMethods({
               {
                 for (var v in inlineFragments
                     .where((frag) => frag.typeCondition != null))
-                  "${v.typeCondition!.on.name.value}": refer(
+                  "${v.typeCondition!.on.name.value}": dataClassAliasMap["${name}__as${v.typeCondition!.on.name.value}"] ?? refer(
                     "${name}__as${v.typeCondition!.on.name.value}",
                   )
               },