… codegen/gql_code_builder/lib/src/operation/data.dart