@knaeckeKami 2023-08-09
Thanks for your PR. I will take a deeper look on the weekend.

IMO a breaking change on this is acceptable for this improvement. If possible without much overhead, I would like to make this configurable though, so users have some time to migrate.