@knaeckeKami 2023-08-12
Looks good to me.

Thanks so much for this contribution.

Also tried it on the codebase I'm working on and seems to work fine (although I don't see much improvement in build times as I don't have many nested, shared fragment spreads ).

In order to ship this, please add a configuration option to the `buildDataLibrary` function which defaults to the current behaviour. 

I suggest adding a simple config class, like for the when() methods for the inline fragment spreads, in order to have the option to add more parameters later without changing the signature. 

The default value when the config variable is not passed should be the current behaviour that emits new classes for every instance of the fragment in order to not break existing users.

From that, I can take the steps to ships this, i.e.:

- add the configuration option to gql_build and ferry_generator
- add end to end codegen tests 
- publish new versions of the affected libraries