@dehypnosis 2023-08-13
Hi @knaeckeKami 

I found and fixed a bug. Also added configuration signature as your explanation. Please check last two commits.
Thanks.

P.S. Quick Question: Is there anyway to deliver options to `gql_build` with `build.yaml` from my app which is using `ferry`? For that, should It be integrated to `ferry_generator`?