@knaeckeKami 2023-08-13
ferry does not use gql_build anymore, as this leads to configuration overhead and user confusion.

But we should still maintain gql_build as other projects depend on it. 

so, this change also has to be integrated into ferry_generator as well.